add esp8266 littleFS support
add new command to handle files/directories
add files creation  sketch for each FS
fix esplogo compilation if no display
remove wifi network starting message if wifi off
