+++ esp3d/configuration.h
@@ -108,8 +108,8 @@
 //FILESYSTEM_FEATURE: to host some files on flash
 //ESP_SPIFFS_FILESYSTEM       0
 //ESP_FAT_FILESYSTEM          1
-//ESP_LITTLEFS_FILESYSTEM     2 //Not Yet implemented
-#define FILESYSTEM_FEATURE ESP_SPIFFS_FILESYSTEM
+//ESP_LITTLEFS_FILESYSTEM     2
+#define FILESYSTEM_FEATURE ESP_FAT_FILESYSTEM
 
 //DIRECT_PIN_FEATURE: allow to access pin using ESP201 command
 #define DIRECT_PIN_FEATURE
