+++ esp3d/src/core/commands.cpp
@@ -453,6 +453,12 @@ bool Commands::execute_internal_command (int cmd, const char* cmd_params, level_
     case 720:
         response = ESP720(cmd_params, auth_type, output);
         break;
+    //Action on ESP Filesystem
+    //rmdir / remove / mkdir / exists
+    //[ESP730]<Action>=<path> pwd=<admin password>
+     case 730:
+        response = ESP730(cmd_params, auth_type, output);
+        break;
 #endif //FILESYSTEM_FEATURE
 
     //Get fw version firmare target and fw version
